Add files via upload
A. What is the important technical information about the dataset that a database administrator would
be interested in? (Hint: Information about the size of the dataset and the nature of the variables)
B. Take a critical look at the data and do a preliminary analysis of the variables. Do a quality check of
the data so that the variables are consistent. Are there any discrepancies present in the data? If yes,
perform preliminary treatment of data.
C. Explore all the features of the data separately by using appropriate visualizations and draw insights
that can be utilized by the business.
D. Understanding the relationships among the variables in the dataset is crucial for every analytical

project. Perform analysis on the data fields to gain deeper insights. Comment on your understanding of
the data. 
E. Employees working on the existing marketing campaign have made the following remarks. Based on
the data and your analysis state whether you agree or disagree with their observations. Justify your
answer Based on the data available.
E1) Steve Roger says “Men prefer SUV by a large margin, compared to the women”
E2) Ned Stark believes that a salaried person is more likely to buy a Sedan.
E3) Sheldon Cooper does not believe any of them; he claims that a salaried male is an easier target for a
SUV sale over a Sedan Sale.
F. From the given data, comment on the amount spent on purchasing automobiles across the following
categories. Comment on how a Business can utilize the results from this exercise. Give justification
along with presenting metrics/charts used for arriving at the conclusions.
Give justification along with presenting metrics/charts used for arriving at the conclusions.
F1) Gender
F2) Personal_loan
G. From the current data set comment if having a working partner leads to the purchase of a higher-
priced car.
H. The main objective of this analysis is to devise an improved marketing strategy to send targeted
information to different groups of potential buyers present in the data. For the current analysis use the
Gender and Marital_status - fields to arrive at groups with similar purchase history.